+++ src/Commands/Commands.cpp
@@ -18,6 +18,8 @@
 #include "DistributionMode.h"
 #include "ShowCurrentInfo.h"
 
+#include "AggressiveStance.h"
+
 DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 {
 	// Load it after Ares'
@@ -43,6 +45,8 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<FPSNewCounterCommandClass>();
 	MakeCommand<DistributionMode1CommandClass>();
 	MakeCommand<DistributionMode2CommandClass>();
+
+	MakeCommand<AggressiveStanceClass>();
 
 	if (Phobos::Config::DevelopmentCommands)
 	{