+++ src/Ext/Techno/Body.h
@@ -78,6 +78,8 @@ public:
 		int LastWarpInDelay;                   // Last-warp in delay for this unit, used by HasCarryoverWarpInDelay.
 		bool IsBeingChronoSphered;             // Set to true on units currently being ChronoSphered, does not apply to Ares-ChronoSphere'd buildings or Chrono reinforcements.
 
+		bool AggressiveStance;                  // Aggressive stance that will auto target buildings
+
 		ExtData(TechnoClass* OwnerObject) : Extension<TechnoClass>(OwnerObject)
 			, TypeExtData { nullptr }
 			, Shield {}
@@ -130,6 +132,7 @@ public:
 			, HasRemainingWarpInDelay { false }
 			, LastWarpInDelay { 0 }
 			, IsBeingChronoSphered { false }
+			, AggressiveStance { false }
 		{ }
 
 		void OnEarlyUpdate();
@@ -168,6 +171,10 @@ public:
 		virtual void InvalidatePointer(void* ptr, bool bRemoved) override;
 		virtual void LoadFromStream(PhobosStreamReader& Stm) override;
 		virtual void SaveToStream(PhobosStreamWriter& Stm) override;
+
+		void InitAggressiveStance();
+		bool GetAggressiveStance() const;
+		void ToggleAggressiveStance();
 
 	private:
 		template <typename T>