+++ src/Ext/Techno/Hooks.cpp
@@ -144,6 +144,8 @@ DEFINE_HOOK(0x6F42F7, TechnoClass_Init, 0x2)
 	if (pExt->TypeExtData->UniqueTechno && pThis->Owner->IsControlledByCurrentPlayer())
 		ScenarioExt::Global()->OwnedHeros.push_back(pExt);
 
+	pExt->InitAggressiveStance();
+
 	return 0;
 }